refactor project di package structure
