+++ app/src/main/java/com/lambdaschool/hackathon_portal/di/activity/ActivityScope.kt
@@ -1,8 +1,8 @@
-package com.lambdaschool.hackathon_portal.di
+package com.lambdaschool.hackathon_portal.di.activity
 
 import javax.inject.Scope
 
 @Scope
 @MustBeDocumented
 @kotlin.annotation.Retention(AnnotationRetention.RUNTIME)
-annotation class FragmentScope
+annotation class ActivityScope
