+++ app/src/main/java/com/lambdaschool/hackathon_portal/di/activity/modules/NavDrawerModule.kt
@@ -1,4 +1,4 @@
-package com.lambdaschool.hackathon_portal.di
+package com.lambdaschool.hackathon_portal.di.activity.modules
 
 import androidx.appcompat.app.ActionBarDrawerToggle
 import androidx.appcompat.widget.Toolbar
@@ -7,6 +7,7 @@ import androidx.navigation.NavController
 import androidx.navigation.Navigation
 import com.google.android.material.navigation.NavigationView
 import com.lambdaschool.hackathon_portal.R
+import com.lambdaschool.hackathon_portal.di.activity.ActivityScope
 import com.lambdaschool.hackathon_portal.ui.MainActivity
 import dagger.Module
 import dagger.Provides
