+++ app/src/main/java/com/lambdaschool/hackathon_portal/di/fragment/FragmentComponent.kt
@@ -1,12 +1,12 @@
-package com.lambdaschool.hackathon_portal.di
+package com.lambdaschool.hackathon_portal.di.fragment
 
 import androidx.fragment.app.Fragment
 import com.lambdaschool.hackathon_portal.ui.fragments.BaseFragment
 import com.lambdaschool.hackathon_portal.ui.fragments.NavDrawerFragment
 import com.lambdaschool.hackathon_portal.ui.fragments.login.LoginFragment
-import com.lambdaschool.hackathon_portal.ui.fragments.login.LoginModule
+import com.lambdaschool.hackathon_portal.di.fragment.modules.LoginModule
 import com.lambdaschool.hackathon_portal.ui.fragments.logout.LogoutFragment
-import com.lambdaschool.hackathon_portal.ui.fragments.logout.LogoutModule
+import com.lambdaschool.hackathon_portal.di.fragment.modules.LogoutModule
 import dagger.BindsInstance
 import dagger.Subcomponent
