+++ app/src/main/java/com/lambdaschool/hackathon_portal/di/fragment/FragmentScope.kt
@@ -1,8 +1,8 @@
-package com.lambdaschool.hackathon_portal.di
+package com.lambdaschool.hackathon_portal.di.fragment
 
 import javax.inject.Scope
 
 @Scope
 @MustBeDocumented
 @kotlin.annotation.Retention(AnnotationRetention.RUNTIME)
-annotation class ActivityScope
+annotation class FragmentScope
